Make MarkerTrackControl a Layout rather than a Control
